requirements.txt
jsonschema>=4.19.0
PyYAML>=6.0.1
pytest>=8.0.0
